feat: Implement mono-API for Vercel Trial
Migrates all endpoints to a single serverless function, resolving 404 errors, improving logging, and staying within Vercel's function limit. Includes comprehensive documentation and configuration updates.
