+++ .env.example
@@ -1,13 +1,16 @@
-# === API Keys ===
-OPENAI_API_KEY=your-openai-api-key
-GOOGLE_MAPS_API_KEY=your-google-maps-api-key
+# Supabase Configuration
 SUPABASE_URL=https://xdhlztmjktminrwmzcpl.supabase.co
-SUPABASE_ANON_KEY=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJzdXBhYmFzZSIsInJlZiI6InhkaGx6dG1qa3RtaW5yd216Y3BsIiwicm9sZSI6ImFub24iLCJpYXQiOjE3NTY3MjgwMTEsImV4cCI6MjA3MjMwNDAxMX0.EmvBqbygr4VLD3PXFaPjbChakRi5YtSrxp8e_K7ZyGY
-SERVICE_ROLE=your-supabase-service-role
-GEMINI_API_KEY=
-# Google Cloud auth JSON file path for Text-to-Speech
-GOOGLE_APPLICATION_CREDENTIALS=./service-account.json
-# === CORS ===
-# Dozwolone domeny frontendu (oddzielone przecinkami)
-# Użyj * żeby dopuścić wszystkie (echo origin)
-CORS_ALLOWED_ORIGINS=http://localhost:5173,https://freeflow-front.vercel.app,https://freeflow-frontend-chi.vercel.app
+SUPABASE_ANON_KEY=your_supabase_anon_key
+SUPABASE_SERVICE_ROLE_KEY=your_supabase_service_role_key
+
+# OpenAI Configuration
+OPENAI_API_KEY=your_openai_api_key
+
+# Google Cloud Configuration (for TTS/ASR)
+GOOGLE_MAPS_API_KEY=your_google_maps_api_key
+GMAPS_KEY=your_google_maps_api_key
+GOOGLE_APPLICATION_CREDENTIALS=path/to/service-account.json
+
+# Server Configuration
+PORT=3003
+HOST=0.0.0.0
