+++ .vercelignore
@@ -1,0 +1,33 @@
+# Test files
+test-*.js
+*.spec.ts
+*.spec.js
+tests/
+
+# Scripts
+scripts/
+
+# Documentation
+*.md
+!README.md
+
+# Development files
+.env
+.env.local
+.env.*.local
+
+# Logs
+*.log
+orders-log.ndjson
+
+# IDE
+.vscode/
+.idea/
+
+# Git
+.git/
+.gitignore
+
+# Other
+GITHUB
+vers
