+++ DOCS_INDEX.md
@@ -1,0 +1,192 @@
+# 📚 Dokumentacja - Spis Treści
+
+## 🚀 Start Tutaj
+
+Jeśli jesteś tutaj po raz pierwszy, zacznij od:
+
+1. **[ROZWIAZANIE.md](./ROZWIAZANIE.md)** - Co zostało naprawione i dlaczego
+2. **[QUICK_START.md](./QUICK_START.md)** - Deploy w 3 krokach
+3. **[PRE_DEPLOY_CHECKLIST.md](./PRE_DEPLOY_CHECKLIST.md)** - Checklist przed deployem
+
+---
+
+## 📖 Główne Dokumenty
+
+### Dla początkujących
+
+| Dokument | Opis | Czas czytania |
+|----------|------|---------------|
+| **[README.md](./README.md)** | Główna dokumentacja projektu | 5 min |
+| **[QUICK_START.md](./QUICK_START.md)** | Szybki start - deploy w 3 krokach | 3 min |
+| **[ROZWIAZANIE.md](./ROZWIAZANIE.md)** | Co zostało naprawione | 5 min |
+| **[FAQ.md](./FAQ.md)** | Najczęściej zadawane pytania | 10 min |
+
+### Dla zaawansowanych
+
+| Dokument | Opis | Czas czytania |
+|----------|------|---------------|
+| **[DEPLOYMENT.md](./DEPLOYMENT.md)** | Pełny przewodnik deployment | 10 min |
+| **[ARCHITECTURE.md](./ARCHITECTURE.md)** | Architektura i diagramy | 10 min |
+| **[MONO_API_INFO.md](./MONO_API_INFO.md)** | Szczegóły mono-API | 5 min |
+| **[PRE_DEPLOY_CHECKLIST.md](./PRE_DEPLOY_CHECKLIST.md)** | Checklist deployment | 5 min |
+
+### Dla developerów
+
+| Dokument | Opis | Czas czytania |
+|----------|------|---------------|
+| **[CHANGELOG.md](./CHANGELOG.md)** | Historia zmian | 3 min |
+| **[.env.example](./.env.example)** | Przykładowe zmienne środowiskowe | 1 min |
+
+---
+
+## 🎯 Dokumenty według zadania
+
+### Chcę zdeployować na Vercel
+1. [QUICK_START.md](./QUICK_START.md) - Szybki start
+2. [PRE_DEPLOY_CHECKLIST.md](./PRE_DEPLOY_CHECKLIST.md) - Checklist
+3. [DEPLOYMENT.md](./DEPLOYMENT.md) - Szczegóły
+
+### Mam problem / błąd
+1. [FAQ.md](./FAQ.md) - Sprawdź FAQ
+2. [ROZWIAZANIE.md](./ROZWIAZANIE.md) - Typowe problemy
+3. [DEPLOYMENT.md](./DEPLOYMENT.md) - Troubleshooting
+
+### Chcę zrozumieć architekturę
+1. [ARCHITECTURE.md](./ARCHITECTURE.md) - Diagramy i flow
+2. [MONO_API_INFO.md](./MONO_API_INFO.md) - Dlaczego mono-API
+3. [api/index.js](./api/index.js) - Kod źródłowy
+
+### Chcę dodać nowy endpoint
+1. [MONO_API_INFO.md](./MONO_API_INFO.md) - Jak dodać endpoint
+2. [api/index.js](./api/index.js) - Edytuj ten plik
+3. [vercel.json](./vercel.json) - Dodaj rewrite
+
+### Chcę uruchomić lokalnie
+1. [README.md](./README.md) - Instalacja
+2. [.env.example](./.env.example) - Zmienne środowiskowe
+3. `npm start` - Uruchom serwer
+
+---
+
+## 📁 Struktura Dokumentacji
+
+```
+docs/
+├── START TUTAJ
+│   ├── ROZWIAZANIE.md          ✅ Co zostało naprawione
+│   ├── QUICK_START.md          🚀 Deploy w 3 krokach
+│   └── PRE_DEPLOY_CHECKLIST.md ✅ Checklist
+│
+├── GŁÓWNE
+│   ├── README.md               📚 Główna dokumentacja
+│   ├── FAQ.md                  ❓ Pytania i odpowiedzi
+│   └── CHANGELOG.md            📝 Historia zmian
+│
+├── TECHNICZNE
+│   ├── DEPLOYMENT.md           📋 Pełny przewodnik deployment
+│   ├── ARCHITECTURE.md         🏗️ Architektura i diagramy
+│   └── MONO_API_INFO.md        💡 Szczegóły mono-API
+│
+└── CONFIG
+    ├── .env.example            ⚙️ Zmienne środowiskowe
+    ├── vercel.json             🔧 Konfiguracja Vercel
+    └── .vercelignore           🚫 Pliki do wykluczenia
+```
+
+---
+
+## 🔍 Szybkie odnośniki
+
+### API Endpoints
+- Health check: `GET /api/health`
+- Lista wszystkich: `GET /api`
+- Dokumentacja: [QUICK_START.md#dostępne-endpointy](./QUICK_START.md)
+
+### Konfiguracja
+- Vercel: [vercel.json](./vercel.json)
+- Environment: [.env.example](./.env.example)
+- Dependencies: [package.json](./package.json)
+
+### Kod źródłowy
+- Mono-API: [api/index.js](./api/index.js)
+- Local server: [server.js](./server.js)
+- Legacy handlers: [lib/handlers/](./lib/handlers/)
+
+### Testy
+- API tests: [tests/api.spec.ts](./tests/api.spec.ts)
+- Smoke tests: [scripts/smoke.mjs](./scripts/smoke.mjs)
+
+---
+
+## 📊 Ścieżki nauki
+
+### Ścieżka 1: Szybki Deploy (15 min)
+```
+ROZWIAZANIE.md → QUICK_START.md → PRE_DEPLOY_CHECKLIST.md → Deploy!
+```
+
+### Ścieżka 2: Pełne zrozumienie (45 min)
+```
+README.md → ARCHITECTURE.md → MONO_API_INFO.md → DEPLOYMENT.md → FAQ.md
+```
+
+### Ścieżka 3: Development (30 min)
+```
+README.md → api/index.js → server.js → tests/ → scripts/
+```
+
+---
+
+## 🆘 Potrzebujesz pomocy?
+
+1. **Sprawdź [FAQ.md](./FAQ.md)** - może tam jest odpowiedź
+2. **Zobacz [ROZWIAZANIE.md](./ROZWIAZANIE.md)** - typowe problemy
+3. **Sprawdź logi** w Vercel Dashboard
+4. **Otwórz Issue** na GitHub
+
+---
+
+## ✅ Quick Reference
+
+### Najważniejsze komendy
+```bash
+# Instalacja
+npm install
+
+# Local dev
+npm start
+
+# Deploy (CLI)
+vercel --prod
+
+# Testy
+npm test
+npm run smoke
+```
+
+### Najważniejsze pliki
+- `api/index.js` - mono-API (wszystkie endpointy)
+- `vercel.json` - konfiguracja Vercel
+- `.env.example` - template zmiennych środowiskowych
+
+### Najważniejsze linki
+- [Vercel Dashboard](https://vercel.com/dashboard)
+- [Supabase Dashboard](https://supabase.com/dashboard)
+- [Google Cloud Console](https://console.cloud.google.com)
+
+---
+
+## 📝 Wkład w dokumentację
+
+Znalazłeś błąd? Coś jest niejasne?
+
+1. Fork repo
+2. Popraw dokumentację
+3. Pull request
+4. Dziękujemy! 🎉
+
+---
+
+**Ostatnia aktualizacja:** 2025-10-01  
+**Wersja:** 1.0.0  
+**Status:** ✅ Gotowe do użycia
